+++ changelog.md
@@ -2,6 +2,7 @@
 
     - New flag `--loop` to loop video indefinitely while streaming. Closes
       #113.
+    - New flag `--mtype` to specify media type. Closes #128.
 
 * mkchromecast (0.3.8.1) **2017/12/24**